+++ usr.bin/f77/libF77/r_atan.c
@@ -3,10 +3,10 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)r_atan.c	5.1	6/7/85
+ *	@(#)r_atan.c	5.2	7/8/85
  */
 
-double r_atan(x)
+float r_atan(x)
 float *x;
 {
 double atan();
