+++ usr.bin/f77/libF77/r_sinh.c
@@ -3,10 +3,10 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)r_sinh.c	5.1	6/7/85
+ *	@(#)r_sinh.c	5.2	7/8/85
  */
 
-double r_sinh(x)
+float r_sinh(x)
 float *x;
 {
 double sinh();
